\name{predict.FRESA_FILTERFIT}
\alias{predict.FRESA_FILTERFIT}
\title{Predicts \code{\link{filteredFit}} models}
\description{
	This function predicts the outcome from a filteredFit model
}
\usage{
	\method{predict}{FRESA_FILTERFIT}(object,...)
}
\arguments{
	\item{object}{
		An object of class FRESA_FILTERFIT 
	}
	\item{...}{
		A list with: testdata=testdata
	}
}
\value{
	the predicted outcome
}
\author{Jose G. Tamez-Pena}
\keyword{Model_Prediction}
